Removes day3 file and adds new partNumber file
PartNumber required for the Day 3 Part 1 problem
