--- day3/day3.go
@@ -1,16 +1,0 @@
-package day3
-
-type PartNumber struct {
-	Number   int
-	IsValid  bool
-	StartIdx int
-	EndIdx   int
-}
-
-type Gear struct {
-	PartNumbers []int
-}
-
-func (gear *Gear) Ratio() int {
-	return gear.PartNumbers[0] * gear.PartNumbers[1]
-}
