+++ day3/partNumber.go
@@ -1,0 +1,8 @@
+package day3
+
+type PartNumber struct {
+	Number   int
+	IsValid  bool
+	StartIdx int
+	EndIdx   int
+}
